requirements.txt
chromadb
torch
click
logging
langchain
huggingface-hub
transformers
InstructorEmbedding
llama-cpp-python
python-docx